extended movie options and bug fixes

+++ R/makeframe.R
@@ -11,10 +11,11 @@
 #' @param frame.index list of frame indices 'x' to be included in the movie
 #' @param width number of pixels along the horizontal axis
 #' @param height number of pixels along the vertical axis
+#' @param pngfile optional path+filename of output file to save the image. R must have write access to this file.
 #'
 #' @author Danail Obreschkow
 #'
-#' @return None
+#' @return Returns the displayed image as n-by-my-by 4 array, representing the 4 RGBA channels of height n and width m.
 #'
 #' @seealso \code{\link{makemovie}}
 #'
@@ -40,15 +41,19 @@
 #'
 #' @export
 
-makeframe = function(frame.draw,frame.index,width=1080,height=720) {
+makeframe = function(frame.draw,frame.index,width=1080,height=720,pngfile=NULL) {
 
   # safe use of par()
   oldpar = par(no.readonly = TRUE)
   on.exit(par(oldpar))
 
-  # make unique frame name
-  fn = tempfile()#sprintf('frame_tmp_%.0f.png',as.numeric(Sys.time())*1e3)
-  on.exit(unlink(fn))
+  # make unique frame file name
+  if (is.null(pngfile)) {
+    fn = tempfile()
+    on.exit(unlink(fn)) # deletes the frame on exit
+  } else {
+    fn = pngfile
+  }
 
   # write frames
   grDevices::png(fn,width=width,height=height)
@@ -60,5 +65,8 @@ makeframe = function(frame.draw,frame.index,width=1080,height=720) {
   par(bg='grey',mar=c(0,0,0,0))
   nplot(xlim=c(0,width),ylim=c(0,height),asp=1)
   rasterImage(img,0,0,width,height)
+
+  # return
+  invisible(img)
 
 }
